+++ tests/sshunit2_test.sh
@@ -75,6 +75,16 @@ testSshunit2ShouldCreateATestFileInSubfolder()
     assertTrue "the test file was generated in the subfolder" "[ -e ${target_dir}/tests/folder/cmd_test.sh ]"
 }
 
+testSshunit2ShouldCreateALibFile()
+{
+    target_dir="${HOME}/myproject"
+    ${SSHUNIT2} -p "${target_dir}" &> /dev/null
+    cd "${target_dir}"
+    ${SSHUNIT2} -l "folder/mylib"
+    assertTrue "the library file was generated" "[ -e ${target_dir}/src/folder/mylib.sh ]"
+    assertFalse "the test file is not executable" "[ -x ${target_dir}/src/folder/mylib.sh ]"
+}
+
 testTheTestFileIsBuiltFromTemplates()
 {
     header=$(cat "${MAINDIR}/src/templates/header.tpl")